pyproject.toml
[project]
name = "capstone-project-unfc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=3.0.0",
    "python-dotenv>=1.2.1",
    "supabase>=2.27.2",
    "yfinance>=1.1.0",
]
